feat: audio controls and processing enhancements
- Added audio controls in `audio-controls.js` with improved event handling.
- Enhanced audio processing capabilities in `audio-processor.js`, including oscillator management and microphone integration.
- Introduced synthesis engines for FM synthesis and sine wave generation in `fm-synthesis.js` and `sine-wave.js`.
- Developed grid mapping functions for musical note generation based on frame data in `circle-of-fifths.js` and `hex-tonnetz.js`.
- Created utility functions for asynchronous error handling in `async.js`.
- Established a new project file generation script in `generate-project-files.sh` to streamline project documentation.
- Added comprehensive logging throughout the audio processing and synthesis modules for better debugging and monitoring.
- Updated state management in `state.js` to include new settings and configurations.
