+++ future/trashcan/generate-project-files.sh
