+++ future/web/audio/synthesis-engines/available-engines.json
@@ -1,1 +1,10 @@
-
+[
+  {
+    "id": "sine-wave",
+    "createdAt": 1750899236911.1191
+  },
+  {
+    "id": "fm-synthesis",
+    "createdAt": 1750899236897.1191
+  }
+]
