+++ future/web/audio/synthesis-engines/sine-wave.js
