+++ future/web/synthesis-grids/available-grids.json
@@ -1,1 +1,11 @@
 
+[
+  {
+    "id": "hex-tonnetz",
+    "createdAt": 1750899236982.1191
+  },
+  {
+    "id": "circle-of-fifths",
+    "createdAt": 1750899236950.1191
+  }
+]
