--- future/web/synthesis-methods/engines/availableEngines.json
@@ -1,10 +1,0 @@
-[
-  {
-    "id": "sine-wave",
-    "createdAt": 1750899236911.1191
-  },
-  {
-    "id": "fm-synthesis",
-    "createdAt": 1750899236897.1191
-  }
-]
