--- future/web/synthesis-methods/grids/availableGrids.json
@@ -1,10 +1,0 @@
-[
-  {
-    "id": "hex-tonnetz",
-    "createdAt": 1750899236982.1191
-  },
-  {
-    "id": "circle-of-fifths",
-    "createdAt": 1750899236950.1191
-  }
-]
