+++ future/web/test/video-capture.test.js
@@ -1,12 +1,54 @@
-describe('processFrame', () => {
-  test('handles non-finite dimensions', async () => {
-    jest.spyOn(global.console, 'error').mockImplementation(() => {});
-    jest.spyOn(structuredLog, 'call').mockImplementation(() => {});
-    const DOM = { frameCanvas: { getContext: () => null }, videoFeed: {} };
-    jest.spyOn(getDOM, 'call').mockReturnValue(DOM);
-    const result = await processFrame(NaN, 240);
-    expect(result).toEqual({ notes: [], newFrameData: null, avgIntensity: 0 });
-    expect(structuredLog).toHaveBeenCalledWith('DEBUG', 'processFrame dimensions', { rawWidth: NaN, rawHeight: 240 });
-    expect(console.error).toHaveBeenCalledWith("Canvas context not found");
+// File: web/test/video-capture.test.js
+import { setupVideoCapture, cleanupVideoCapture } from '../ui/video-capture.js';
+import { structuredLog } from '../utils/logging.js';
+import { getDOM } from '../core/context.js';
+import { dispatchEvent } from '../core/dispatcher.js';
+
+jest.mock('../utils/logging.js', () => ({
+  structuredLog: jest.fn()
+}));
+jest.mock('../core/context.js', () => ({
+  getDOM: jest.fn()
+}));
+jest.mock('../core/dispatcher.js', () => ({
+  dispatchEvent: jest.fn()
+}));
+
+describe('video-capture', () => {
+  test('setupVideoCapture handles missing DOM elements', async () => {
+    const DOM = { videoFeed: null, frameCanvas: null };
+    const result = await setupVideoCapture(DOM);
+    expect(result).toBe(false);
+    expect(structuredLog).toHaveBeenCalledWith('ERROR', 'Missing videoFeed or frameCanvas in setupVideoCapture');
+    expect(dispatchEvent).toHaveBeenCalledWith('logError', { message: 'Missing videoFeed or frameCanvas in setupVideoCapture' });
+  });
+
+  test('setupVideoCapture initializes video feed and canvas', async () => {
+    const DOM = {
+      videoFeed: { setAttribute: jest.fn() },
+      frameCanvas: { style: { display: '' }, setAttribute: jest.fn() }
+    };
+    const result = await setupVideoCapture(DOM);
+    expect(result).toBe(true);
+    expect(DOM.videoFeed.setAttribute).toHaveBeenCalledWith('autoplay', '');
+    expect(DOM.videoFeed.setAttribute).toHaveBeenCalledWith('muted', '');
+    expect(DOM.videoFeed.setAttribute).toHaveBeenCalledWith('playsinline', '');
+    expect(DOM.frameCanvas.style.display).toBe('none');
+    expect(DOM.frameCanvas.setAttribute).toHaveBeenCalledWith('aria-hidden', 'true');
+    expect(structuredLog).toHaveBeenCalledWith('INFO', 'setupVideoCapture: Video feed and canvas initialized');
+  });
+
+  test('cleanupVideoCapture clears video feed and canvas', async () => {
+    const DOM = {
+      videoFeed: { srcObject: { getTracks: () => [{ stop: jest.fn() }] }, srcObject: null },
+      frameCanvas: { width: 0, height: 0 }
+    };
+    getDOM.mockReturnValue(DOM);
+    await cleanupVideoCapture();
+    expect(DOM.videoFeed.srcObject.getTracks()[0].stop).toHaveBeenCalled();
+    expect(DOM.videoFeed.srcObject).toBe(null);
+    expect(DOM.frameCanvas.width).toBe(0);
+    expect(DOM.frameCanvas.height).toBe(0);
+    expect(structuredLog).toHaveBeenCalledWith('INFO', 'cleanupVideoCapture: Video capture cleaned up');
   });
 });
