+++ future/web/ui/cleanup-manager.js
@@ -1,6 +1,5 @@
-// ui/cleanup-manager.js
-import { settings, setStream, setAudioInterval } from "../state.js";
-import { cleanupAudio } from "../audio-processor.js";
+import { settings, setStream, setAudioInterval } from '../core/state.js';
+import { cleanupAudio } from '../audio/audio-processor.js';
 
 let isAudioInitialized = false;
 let audioContext = null;
